requirements.txt
torch==0.4.1
torchvision==0.2.1
tensorboardX
pillow
numpy
opencv-contrib-python

# API Server dependencies
fastapi==0.61.1
python-multipart==0.0.5
aiofiles==0.6.0
